// model/profile.ts
export type ProfileRequest = {
    Email: string
    Username: string
    Password: string
    Name: string
    Phone: string
    Address: string
}

export type MerchantProfile = {
    Email: string
    Username: string
    Password: string
    Name: string
    Phone: string
    Address: string
    Latitude: string
    Longitude: string
}